backend/requirements.txt
bidict==0.22.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.0.4
colorama==0.4.6
cryptography==41.0.3
dnspython==2.4.2
eventlet==0.33.3
filelock==3.12.3
Flask==2.2.2
Flask-Cors==4.0.0
Flask-SocketIO==5.3.6
future==0.18.3
greenlet==2.0.2
h11==0.14.0
idna==3.4
itsdangerous==2.0.1
Jinja2==3.1.2
MarkupSafe==2.1.1
ndg-httpsclient==0.5.1
pip==23.2.1
pyasn1==0.5.0
pycparser==2.21
pyOpenSSL==23.2.0
PySocks==1.7.1
python-dotenv==1.0.0
python-engineio==4.7.0
python-socketio==5.9.0
requests==2.31.0
requests-file==1.5.1
setuptools==68.0.0
simple-websocket==0.10.1
six==1.16.0
tldextract==3.5.0
urllib3==2.0.4
Werkzeug==2.2.3
wheel==0.38.4
wsproto==1.2.0
zope.event==5.0
zope.interface==6.0
